requirements.txt
ftfy
regex
tqdm
torch
torchvision
torch>=1.6.0
miditoolkit
PyYAML
scipy
numpy
matplotlib
pylab-sdk
requests